Big restructuring to make things more clear
