+++ Dataset/EuclideanClusteredData/Cluster2.csv
@@ -1,86 +1,23 @@
 ID#,Balance,Qual_miles,cc1_miles,cc2_miles,cc3_miles,Bonus_miles,Bonus_trans,Flight_miles_12mo,Flight_trans_12,Days_since_enroll,Award?
-33,276571,0,4,1,1,42044,23,0,0,7872,0
-46,288865,967,1,1,1,23600,14,2000,4,6884,1
-64,362642,0,1,1,1,28079,8,0,0,6835,0
-114,298911,0,1,1,1,8550,11,2550,8,7714,1
-117,295638,0,1,1,1,5422,20,818,2,7700,1
-155,298121,0,5,1,1,60808,20,500,1,7640,0
-171,370941,0,4,1,1,44615,16,0,0,7577,1
-175,352508,0,3,1,1,23740,6,1000,2,7563,1
-196,386061,0,5,1,1,107813,24,0,0,7500,1
-241,285116,0,1,1,1,8471,18,500,1,7406,1
-296,294881,0,5,1,1,84800,20,0,0,7279,1
-307,364387,0,4,1,1,28200,27,0,0,7213,0
-311,383030,2998,1,1,1,8001,25,3226,15,7326,1
-327,278727,0,1,1,1,21666,19,9166,16,7155,1
-331,388455,0,5,1,1,114329,26,6078,8,7164,1
-340,278457,6727,4,1,1,57313,27,1000,2,7101,1
-357,312512,0,1,1,1,100,1,100,1,7078,0
-385,377252,0,5,1,1,230629,30,6393,13,6996,1
-443,353678,0,4,1,1,141230,28,1450,5,8289,1
-454,383129,0,3,1,1,23608,16,500,1,8289,0
-471,382546,0,5,1,1,136619,34,3000,6,8228,1
-521,346178,0,5,1,1,63803,16,0,0,6645,0
-540,267404,0,5,1,1,263685,51,3000,6,6584,1
-568,380930,0,2,2,1,22485,25,213,1,6539,1
-607,329497,0,5,1,1,57264,12,0,0,6484,1
-627,338656,0,4,1,1,52908,16,0,0,6444,0
-633,365918,0,5,1,1,83752,23,1373,7,6429,1
-658,347096,0,1,1,1,12000,6,0,0,6408,1
-667,323459,0,1,1,1,12000,8,1500,3,6378,1
-688,280561,0,3,1,1,12154,13,0,0,6337,0
-693,369816,1745,4,1,1,46942,35,7266,17,6340,1
-712,340874,0,5,1,1,68090,43,5676,27,6305,1
-733,384470,0,3,1,1,20468,16,0,0,6280,1
-791,333500,0,5,1,1,69780,23,1258,4,6171,1
-797,293648,0,3,1,1,22148,15,0,0,6161,0
-830,363165,2565,1,1,1,5950,11,1850,8,6116,1
-842,292533,0,5,1,1,64186,15,1000,2,6095,1
-880,323273,0,1,1,1,9000,5,500,1,6028,1
-891,301447,0,5,1,1,63866,18,500,1,6004,1
-916,289351,0,5,1,1,147787,35,500,1,5789,1
-922,322454,0,4,1,1,44591,23,250,1,5961,1
-943,290012,0,5,1,1,64360,18,650,2,5931,1
-974,281626,0,5,1,1,53521,16,0,0,5878,0
-1010,370596,0,5,1,1,50120,12,0,0,5826,0
-1013,330644,0,1,1,1,0,0,0,0,5822,0
-1069,305519,0,5,1,1,79216,27,650,3,5724,1
-1075,354641,0,5,1,1,77694,29,3600,9,5710,1
-1123,295008,0,5,1,1,136704,27,0,0,5028,1
-1138,375927,0,2,1,1,18202,30,1100,3,5652,1
-1232,378759,0,1,1,1,28904,9,2303,4,5408,1
-1327,286101,1000,5,1,1,72965,31,3000,11,4666,1
-1398,320997,2556,5,1,1,82255,24,0,0,5345,1
-1411,355145,0,5,1,1,61926,29,806,8,5288,0
-1473,315273,0,1,1,1,12584,9,4084,4,5079,1
-1561,319241,0,5,1,1,71867,36,0,0,4692,0
-1562,333346,2139,4,1,1,52013,31,4309,12,4469,1
-1620,278437,0,4,1,1,50888,26,0,0,4744,0
-1635,276607,0,1,1,1,11430,14,6430,11,4791,1
-1640,307260,783,4,1,1,40533,24,250,1,4791,1
-1654,385195,0,1,1,1,0,0,0,0,4903,0
-1685,288379,0,1,1,1,15942,8,1442,5,4741,1
-1719,297127,0,3,1,1,27415,16,0,0,4631,1
-1857,368535,0,5,1,1,146201,21,2600,6,4210,1
-1889,264685,0,5,1,1,57505,16,0,0,4112,0
-1924,355423,5009,3,1,1,20500,20,100,1,3963,1
-1927,283670,0,4,1,1,39260,39,200,1,3972,0
-2016,327126,1452,5,1,1,97997,25,350,3,4440,1
-2018,268097,0,3,1,1,26933,25,0,0,4105,0
-2021,354807,0,1,1,1,6000,12,3500,11,4597,1
-2066,379744,2081,1,1,1,19191,17,1850,11,4608,1
-2189,268314,0,4,1,1,29968,14,500,1,3869,0
-2232,292023,0,1,1,1,14125,12,3250,7,3683,1
-2361,350608,0,4,1,1,50988,26,2643,5,3630,0
-2491,307149,2562,4,1,1,58008,19,650,2,3173,1
-2545,326704,0,4,1,1,37408,19,0,0,2962,0
-2627,273442,0,4,1,1,54625,35,2850,7,2676,0
-2693,282784,3508,1,1,1,19314,17,3100,7,2460,1
-2746,278258,0,1,1,1,1000,1,0,0,2298,0
-2782,282971,0,1,1,1,5780,16,300,3,2192,1
-2960,310226,0,4,1,1,61894,28,1700,9,2900,1
-3006,326201,0,1,1,1,19350,14,0,0,3119,0
-3159,298505,0,5,1,1,130647,32,600,2,2483,1
-3257,287033,0,1,1,1,26161,58,12873,53,2272,1
-3338,298927,0,5,1,1,69449,27,150,3,2047,0
-3811,374613,0,1,2,1,23592,31,5000,14,3463,1
+44,619393,0,3,1,1,15008,14,0,0,7819,1
+90,609477,0,3,1,1,21422,22,1200,8,6820,0
+120,969559,0,1,1,1,2500,3,1000,2,7718,1
+129,602064,0,5,1,1,194753,26,2250,10,7652,1
+154,930410,0,5,1,1,211284,18,2250,6,7640,1
+190,707079,0,4,1,1,57173,40,3450,12,7523,1
+294,568174,0,5,1,1,67121,16,1000,2,7243,1
+317,714717,0,5,1,1,119162,20,1750,4,7183,1
+322,766419,0,1,1,1,11398,3,398,1,7162,1
+416,620498,0,5,1,1,25395,53,1750,14,8296,1
+477,743745,2066,1,1,1,4100,20,2500,18,8228,1
+516,889456,1024,5,1,1,98959,29,850,4,6652,1
+638,747983,0,4,1,1,26117,15,50,1,6427,1
+833,765738,0,5,1,1,62133,14,0,0,6112,1
+1127,739476,0,1,1,1,24215,8,322,1,5678,1
+1200,837979,0,1,1,1,18372,15,1450,4,5373,1
+1313,737514,0,3,1,4,43993,40,2150,7,5030,0
+1672,642756,0,4,1,1,37889,15,1000,2,4841,0
+1956,640169,0,5,1,1,82561,36,400,2,3895,1
+2508,704113,0,5,1,1,98119,23,0,0,3082,1
+2676,644881,2600,1,1,1,14350,28,10850,26,2511,1
+3795,822321,0,5,1,1,138334,26,600,2,3288,1
