+++ Dataset/EuclideanClusteredData/Cluster5.csv
@@ -1,6 +1,90 @@
 ID#,Balance,Qual_miles,cc1_miles,cc2_miles,cc3_miles,Bonus_miles,Bonus_trans,Flight_miles_12mo,Flight_trans_12,Days_since_enroll,Award?
-224,1302051,2706,5,1,1,90653,32,3050,7,7467,1
-279,1704838,0,1,1,1,17108,32,4823,23,7283,1
-472,1113394,2015,5,1,1,60883,26,2600,9,8228,1
-487,1125076,0,3,1,1,23046,14,0,0,8107,0
-494,1199220,2919,1,1,1,3800,19,3300,18,8023,1
+9,443003,0,3,2,1,1753,43,3850,12,6948,1
+64,362642,0,1,1,1,28079,8,0,0,6835,0
+74,550367,0,3,1,1,12500,13,50,1,7801,0
+170,402874,0,5,1,1,74800,15,0,0,7582,1
+171,370941,0,4,1,1,44615,16,0,0,7577,1
+175,352508,0,3,1,1,23740,6,1000,2,7563,1
+178,451673,0,4,1,1,43533,19,900,4,7575,1
+192,410795,0,5,1,1,73679,31,0,0,7514,0
+196,386061,0,5,1,1,107813,24,0,0,7500,1
+248,468175,0,5,1,1,141615,22,0,0,7348,1
+260,423540,0,1,1,1,8534,14,2100,9,7375,0
+263,402312,0,5,1,1,77122,18,250,3,7375,1
+307,364387,0,4,1,1,28200,27,0,0,7213,0
+311,383030,2998,1,1,1,8001,25,3226,15,7326,1
+329,455228,0,1,1,1,258,2,258,2,7141,1
+331,388455,0,5,1,1,114329,26,6078,8,7164,1
+357,312512,0,1,1,1,100,1,100,1,7078,0
+385,377252,0,5,1,1,230629,30,6393,13,6996,1
+388,479989,0,4,1,1,66516,26,1600,6,6980,0
+410,507506,1952,5,1,1,102279,32,4485,13,8296,1
+426,529886,7210,1,1,1,23866,26,7741,15,8296,1
+433,491987,0,5,1,1,68007,23,700,2,8289,1
+443,353678,0,4,1,1,141230,28,1450,5,8289,1
+454,383129,0,3,1,1,23608,16,500,1,8289,0
+471,382546,0,5,1,1,136619,34,3000,6,8228,1
+473,459681,0,5,1,1,95531,17,0,0,8228,1
+521,346178,0,5,1,1,63803,16,0,0,6645,0
+536,413509,0,1,1,1,2500,1,0,0,6598,1
+568,380930,0,2,2,1,22485,25,213,1,6539,1
+607,329497,0,5,1,1,57264,12,0,0,6484,1
+616,427528,0,4,1,1,35098,16,1600,3,6465,0
+627,338656,0,4,1,1,52908,16,0,0,6444,0
+633,365918,0,5,1,1,83752,23,1373,7,6429,1
+639,415070,588,5,1,1,153183,41,3802,20,6423,1
+658,347096,0,1,1,1,12000,6,0,0,6408,1
+667,323459,0,1,1,1,12000,8,1500,3,6378,1
+693,369816,1745,4,1,1,46942,35,7266,17,6340,1
+703,463494,0,3,1,1,27390,17,750,1,6323,1
+707,529477,0,2,1,1,31624,28,2098,5,6311,1
+711,417875,0,4,1,1,44472,18,0,0,6308,0
+712,340874,0,5,1,1,68090,43,5676,27,6305,1
+733,384470,0,3,1,1,20468,16,0,0,6280,1
+753,440265,4081,1,1,1,22750,15,1250,10,6234,1
+791,333500,0,5,1,1,69780,23,1258,4,6171,1
+830,363165,2565,1,1,1,5950,11,1850,8,6116,1
+880,323273,0,1,1,1,9000,5,500,1,6028,1
+891,301447,0,5,1,1,63866,18,500,1,6004,1
+899,444159,0,1,1,1,0,0,0,0,5993,0
+922,322454,0,4,1,1,44591,23,250,1,5961,1
+1010,370596,0,5,1,1,50120,12,0,0,5826,0
+1013,330644,0,1,1,1,0,0,0,0,5822,0
+1069,305519,0,5,1,1,79216,27,650,3,5724,1
+1075,354641,0,5,1,1,77694,29,3600,9,5710,1
+1096,427357,0,2,1,1,6027,12,0,0,5587,1
+1123,295008,0,5,1,1,136704,27,0,0,5028,1
+1138,375927,0,2,1,1,18202,30,1100,3,5652,1
+1232,378759,0,1,1,1,28904,9,2303,4,5408,1
+1398,320997,2556,5,1,1,82255,24,0,0,5345,1
+1411,355145,0,5,1,1,61926,29,806,8,5288,0
+1442,472241,500,1,1,1,3600,12,3500,11,5045,1
+1473,315273,0,1,1,1,12584,9,4084,4,5079,1
+1518,507371,0,4,1,1,51765,18,450,1,4770,1
+1520,391222,0,3,1,1,27471,20,1600,7,4905,1
+1561,319241,0,5,1,1,71867,36,0,0,4692,0
+1562,333346,2139,4,1,1,52013,31,4309,12,4469,1
+1617,404142,0,5,1,1,75509,13,0,0,4784,1
+1640,307260,783,4,1,1,40533,24,250,1,4791,1
+1649,533120,0,5,1,1,120983,18,0,0,4827,1
+1654,385195,0,1,1,1,0,0,0,0,4903,0
+1781,431824,0,3,1,1,30480,16,500,1,4443,1
+1857,368535,0,5,1,1,146201,21,2600,6,4210,1
+1860,456509,0,1,1,1,500,1,500,1,4201,0
+1924,355423,5009,3,1,1,20500,20,100,1,3963,1
+2016,327126,1452,5,1,1,97997,25,350,3,4440,1
+2021,354807,0,1,1,1,6000,12,3500,11,4597,1
+2066,379744,2081,1,1,1,19191,17,1850,11,4608,1
+2087,475338,0,4,1,1,39680,22,700,4,3355,0
+2157,394999,0,1,1,1,4150,5,150,3,3599,0
+2361,350608,0,4,1,1,50988,26,2643,5,3630,0
+2447,451294,0,3,1,1,45907,22,500,1,3347,1
+2491,307149,2562,4,1,1,58008,19,650,2,3173,1
+2545,326704,0,4,1,1,37408,19,0,0,2962,0
+2583,518960,0,1,1,1,15700,8,200,1,2819,1
+2731,405499,1000,4,1,1,72318,46,9729,29,2329,1
+2960,310226,0,4,1,1,61894,28,1700,9,2900,1
+3006,326201,0,1,1,1,19350,14,0,0,3119,0
+3062,449205,1121,1,1,1,4750,9,3250,8,3071,1
+3159,298505,0,5,1,1,130647,32,600,2,2483,1
+3811,374613,0,1,2,1,23592,31,5000,14,3463,1
