+++ Dataset/EuclideanClusteredData/Cluster6.csv
@@ -1,12 +1,6 @@
 ID#,Balance,Qual_miles,cc1_miles,cc2_miles,cc3_miles,Bonus_miles,Bonus_trans,Flight_miles_12mo,Flight_trans_12,Days_since_enroll,Award?
-120,969559,0,1,1,1,2500,3,1000,2,7718,1
-154,930410,0,5,1,1,211284,18,2250,6,7640,1
-322,766419,0,1,1,1,11398,3,398,1,7162,1
-477,743745,2066,1,1,1,4100,20,2500,18,8228,1
-516,889456,1024,5,1,1,98959,29,850,4,6652,1
-638,747983,0,4,1,1,26117,15,50,1,6427,1
-833,765738,0,5,1,1,62133,14,0,0,6112,1
-1127,739476,0,1,1,1,24215,8,322,1,5678,1
-1200,837979,0,1,1,1,18372,15,1450,4,5373,1
-1313,737514,0,3,1,4,43993,40,2150,7,5030,0
-3795,822321,0,5,1,1,138334,26,600,2,3288,1
+224,1302051,2706,5,1,1,90653,32,3050,7,7467,1
+279,1704838,0,1,1,1,17108,32,4823,23,7283,1
+472,1113394,2015,5,1,1,60883,26,2600,9,8228,1
+487,1125076,0,3,1,1,23046,14,0,0,8107,0
+494,1199220,2919,1,1,1,3800,19,3300,18,8023,1
