+++ Question1/1a/KMeansClusteringEuclidean.py
