+++ Question1/1a/Tools.py
